add `merge` and `static` solver (#27)
the new `merge` solver is able to assemble an IP address from multiple
other solvers, using netmasks to combine them into a single address.
This is especially useful in situations where different parts of the
external IP address have to be retrieved from different sources,
such as when running NPTv6 with different internal and external prefixes.

Meanwhile, the `static` solver can serve as a fallback for regular solving
and can augment the `merge` solver by providing static parts for the
final address.